Did some slight package refactoring
