+++ multichat/src/main/java/xyz/olivermartin/multichat/local/platform/spigot/MultiChatLocalSpigotPlugin.java
@@ -12,9 +12,9 @@ import xyz.olivermartin.multichat.local.LocalChatManager;
 import xyz.olivermartin.multichat.local.LocalConsoleLogger;
 import xyz.olivermartin.multichat.local.LocalMetaManager;
 import xyz.olivermartin.multichat.local.LocalPlaceholderManager;
+import xyz.olivermartin.multichat.local.LocalProxyCommunicationManager;
 import xyz.olivermartin.multichat.local.MultiChatLocal;
 import xyz.olivermartin.multichat.local.MultiChatLocalPlatform;
-import xyz.olivermartin.multichat.local.communication.LocalProxyCommunicationManager;
 import xyz.olivermartin.multichat.local.config.LocalConfigManager;
 import xyz.olivermartin.multichat.local.platform.spigot.commands.MultiChatLocalSpigotCommand;
 import xyz.olivermartin.multichat.local.platform.spigot.commands.SpigotNickCommand;
