+++ multichat/src/main/java/xyz/olivermartin/multichat/local/platform/spigot/SpigotBungeeCommunicationManager.java
@@ -8,9 +8,9 @@ import java.util.UUID;
 import org.bukkit.Bukkit;
 import org.bukkit.plugin.messaging.PluginMessageRecipient;
 
+import xyz.olivermartin.multichat.local.LocalBungeeCommunicationManager;
 import xyz.olivermartin.multichat.local.MultiChatLocal;
 import xyz.olivermartin.multichat.local.MultiChatLocalPlatform;
-import xyz.olivermartin.multichat.local.communication.LocalBungeeCommunicationManager;
 
 /**
  * Allows MultiChatLocal running on Spigot to communicate with a Bungeecord Proxy
