+++ multichat/src/main/java/xyz/olivermartin/multichat/local/platform/sponge/MultiChatLocalSpongePlugin.java
@@ -21,9 +21,9 @@ import xyz.olivermartin.multichat.local.LocalChatManager;
 import xyz.olivermartin.multichat.local.LocalConsoleLogger;
 import xyz.olivermartin.multichat.local.LocalMetaManager;
 import xyz.olivermartin.multichat.local.LocalPlaceholderManager;
+import xyz.olivermartin.multichat.local.LocalProxyCommunicationManager;
 import xyz.olivermartin.multichat.local.MultiChatLocal;
 import xyz.olivermartin.multichat.local.MultiChatLocalPlatform;
-import xyz.olivermartin.multichat.local.communication.LocalProxyCommunicationManager;
 import xyz.olivermartin.multichat.local.config.LocalConfigManager;
 import xyz.olivermartin.multichat.local.platform.sponge.commands.MultiChatLocalSpongeCommand;
 import xyz.olivermartin.multichat.local.platform.sponge.commands.SpongeNickCommand;
