+++ multichat/src/main/java/xyz/olivermartin/multichat/local/platform/sponge/SpongeBungeeCommunicationManager.java
@@ -8,8 +8,8 @@ import org.spongepowered.api.Sponge;
 import org.spongepowered.api.entity.living.player.Player;
 import org.spongepowered.api.network.ChannelBinding.RawDataChannel;
 
+import xyz.olivermartin.multichat.local.LocalBungeeCommunicationManager;
 import xyz.olivermartin.multichat.local.MultiChatLocalPlatform;
-import xyz.olivermartin.multichat.local.communication.LocalBungeeCommunicationManager;
 
 /**
  * Allows MultiChatLocal running on Sponge to communicate with a Bungeecord Proxy
